requirements.txt
Flask==3.1.2
google-generativeai==0.8.5
gunicorn==23.0.0
python-dotenv==1.1.1
chromadb==1.2.1
hf_xet==1.2.0
pillow==10.0.0
PyMuPDF==1.23.8
numpy<2
